updated the Habit COntext to be named after convex
