+++ src/components/HabitContext.tsx
@@ -1,15 +1,36 @@
 import { createContext } from 'react';
-import { HabitData } from './types';
 import { v4 } from 'uuid';
-import { useSearchParams } from 'react-router-dom';
 import { useState } from 'react';
-import { Habit } from './types';
-import { HabitDictionary } from './types';
 import { useMutation, useQuery } from '../../convex/_generated/react';
 
+export type Habit = {
+  id: string, 
+  when: string, 
+  insteadOf: string, 
+  will : string,
+  _id?: string
+};
+export type HabitDictionary = {
+  [key: string]: Habit;
+}
+export type HabitData = {
+  data: Habit[];
+  currentWhen: string;
+  currentInstead: string;
+  currentWill: string;
+  currentId: string;
+  add: (item?:Habit) => void;
+  remove: (habit:Habit) => void;
+  get: (id:string) => Habit;
+  setCurrentWhen: (id:string) => void;
+  setCurrentInstead: (id:string) => void;
+  setCurrentWill: (id:string) => void;
+  setCurrent: (id?:string) => void;
+  
+}
 const HabitContext = createContext({} as HabitData);
 
-const useHabitContextControler = (cohort:string) => {
+const useConvexHabitContextController = (cohort:string) => {
   const tasks = useQuery("getHabits", {cohort}) as unknown as Habit[];
   const addTask = useMutation("addHabit");
   const delHabit = useMutation("delHabit");
@@ -20,7 +41,6 @@ const useHabitContextControler = (cohort:string) => {
   let [currentWill, setCurrentWill] = useState('');
   let [currentId, setCurrentId] = useState(v4());
   let [current_id, setCurrent_id] = useState('' as string);
-  let [habitId, setHabitId] = useState('');
   const get = (id:string) =>{
       return (tasks || []).find((item) => item.id === id) ||  {when:'', insteadOf:'', will:'', id };
   };
@@ -37,18 +57,11 @@ const useHabitContextControler = (cohort:string) => {
       const item = itemIn || [currentWhen, currentInstead, currentWill, currentId];
       console.log({id:currentId, when:currentWhen, insteadOf:currentInstead, will:currentWill});
       addTask({id:currentId, when:currentWhen, insteadOf:currentInstead, will:currentWill, cohort, _id:current_id});
-      /*
-      setData(Object.values(
-        { 
-          ...data.reduce((acc, cur) => {acc[cur[3]] = cur; return acc}, {} as HabitDictionary), 
-          [item[3]]: item
-        }));
-        */
+    },
+    remove: (record:Habit) => {
+      delHabit(record);
     },
     get,
-    setId: (id:string) => {  
-      setHabitId(id);
-    },
     setCurrent: (id?:string) => {
       console.log({id})
       const {when, insteadOf, will, _id} = typeof id !=='undefined' ? get(id) : {when:'', insteadOf:'', will:'', _id:null}
@@ -59,13 +72,9 @@ const useHabitContextControler = (cohort:string) => {
       setCurrentId(id || v4());
       setCurrent_id((_id || ''));
     },
-    remove: (record:Habit) => {
-      delHabit(record);
-      //setData(data.filter((item) => item[3] !== id));
-    }
   }
   return controller as HabitData;
 }
 
 export default HabitContext;
-export {useHabitContextControler};
+export {useConvexHabitContextController};
